requirements.txt
# 適宜追記
requests
beautifulsoup4
selenium
notion-client
python-dotenv
typing